+++ src/compas_view2/scene/camera.py
@@ -1,18 +1,21 @@
-from compas.geometry import Translation
-from compas.geometry import Rotation
-from compas.geometry import Vector
-from numpy.linalg import norm
-from numpy.linalg import det
 from math import atan2
-from numpy import pi
+from typing import List
+
 from numpy import array
 from numpy import asfortranarray
 from numpy import dot
 from numpy import float32
-from compas_view2.objects import Object
-from typing import List
+from numpy import pi
+from numpy.linalg import det
+from numpy.linalg import norm
 
-from .matrices import perspective, ortho
+from compas.geometry import Rotation
+from compas.geometry import Translation
+from compas.geometry import Vector
+from compas_view2.objects import Object
+
+from .matrices import ortho
+from .matrices import perspective
 
 
 class Position(Vector):
@@ -134,7 +137,7 @@ class Camera:
         self.reset_position()
         if target:
             self.target = target
-        if position:
+        if position and view._current == 4:
             self.position = position
 
     @property
@@ -235,13 +238,13 @@ class Camera:
     def reset_position(self):
         """Reset the position of the camera based current view type."""
         self.target.set(0, 0, 0)
-        if self.view.current == self.view.PERSPECTIVE:
+        if self.view.current == self.view.VIEWPORTS["perspective"]:
             self.rotation.set(pi / 4, 0, -pi / 4)
-        if self.view.current == self.view.TOP:
+        if self.view.current == self.view.VIEWPORTS["top"]:
             self.rotation.set(0, 0, 0)
-        if self.view.current == self.view.FRONT:
+        if self.view.current == self.view.VIEWPORTS["front"]:
             self.rotation.set(pi / 2, 0, 0)
-        if self.view.current == self.view.RIGHT:
+        if self.view.current == self.view.VIEWPORTS["right"]:
             self.rotation.set(pi / 2, 0, pi / 2)
 
     def rotate(self, dx, dy):
@@ -260,10 +263,10 @@ class Camera:
 
         Notes
         -----
-        Camera rotations are only available if the current view is a perspective view (``camera.view.current == camera.view.PERSPECTIVE``).
+        Camera rotations are only available if the current view is a perspective view (``camera.view.current == camera.view.VIEWPORTS["perspective"]``).
 
         """
-        if self.view.current == self.view.PERSPECTIVE:
+        if self.view.current == self.view.VIEWPORTS["perspective"]:
             self.rotation += [-self.rotation_delta * dy, 0, -self.rotation_delta * dx]
 
     def pan(self, dx, dy):
@@ -323,7 +326,7 @@ class Camera:
 
         """
         aspect = width / height
-        if self.view.current == self.view.PERSPECTIVE:
+        if self.view.current == self.view.VIEWPORTS["perspective"]:
             P = perspective(self.fov, aspect, self.near * self.scale, self.far * self.scale)
         else:
             left = -self.distance